REF: Remove BotMaestroSDKInterface and support for Maestro V1 API.
