--- botcity/maestro/impl/__init__.py
@@ -1,9 +1,0 @@
-from .interface import (BotMaestroSDKInterface, ensure_access_token,
-                        ensure_implementation, since_version)
-
-from . import v1
-from . import v2
-
-__all__ = [
-    'BotMaestroSDKInterface', 'ensure_access_token', 'since_version', 'ensure_implementation', 'v1', 'v2'
-]
