+++ botcity/maestro/sdk.py
@@ -1,18 +1,117 @@
-from typing import Dict, List, Optional, Tuple
+import datetime
+import json
+import os
+import platform
+import sys
+import traceback
+import warnings
+from dataclasses import asdict
+from functools import wraps
+from io import IOBase, StringIO
+from pathlib import Path
+from typing import Any, Callable, Dict, List, Optional, Tuple, TypeVar, cast
 
+import distro
+import importlib_metadata
 import requests
 import urllib3
+from packaging import version
+from requests_toolbelt import MultipartEncoder
 
 from . import model
-from .impl import (BotMaestroSDKInterface, v1, v2,
-                   ensure_access_token, ensure_implementation, since_version)
-
 from .datapool import DataPool
 
 urllib3.disable_warnings(urllib3.exceptions.InsecureRequestWarning)
 
 
-class BotMaestroSDK(BotMaestroSDKInterface):
+F = TypeVar('F', bound=Callable[..., Any])
+
+
+def ensure_access_token(invoke: Optional[bool] = False) -> Callable[[F], F]:
+    """
+    Decorator to ensure that a token is available.
+
+    Args:
+        func (callable): The function to be wrapped
+        invoke (bool): Whether or not to invoke the function anyway.
+    Returns:
+        wrapper (callable): The decorated function
+    """
+    def decorator(func: F) -> F:
+        @wraps(func)
+        def wrapper(obj, *args, **kwargs):
+            if isinstance(obj, BotMaestroSDK):
+                if obj.access_token is None:
+                    if obj.RAISE_NOT_CONNECTED:
+                        raise RuntimeError('Access Token not available. Make sure to invoke login first.')
+                    else:
+                        message = ""
+                        if not obj._notified_disconnect:
+                            obj._notified_disconnect = True
+                            message += "** WARNING BotCity Maestro is not logged in and RAISE_NOT_CONNECTED is "
+                            message += "False. Running on Offline mode. **"
+                            warnings.warn(message, stacklevel=2)
+                        message = f"Invoked '{func.__name__}'"
+                        params: List[str] = []
+                        if args:
+                            for ar in args:
+                                params.append(str(ar))
+                        if kwargs:
+                            for k, v in kwargs.items():
+                                params.append(f"{k}={v}")
+                        if params:
+                            message += ' with arguments '
+                            message += ", ".join(params)
+                        message += "."
+                        warnings.warn(message, stacklevel=2)
+                        if not invoke:
+                            return lambda *args, **kwargs: None
+            else:
+                raise NotImplementedError('ensure_token is only valid for BotMaestroSDK methods.')
+            return func(obj, *args, **kwargs)
+
+        return cast(F, wrapper)
+    return decorator
+
+
+def since_version(v: str) -> Callable[[F], F]:
+    """
+    Decorator to ensure that a method is availble for a given Maestro backend version.
+
+    Args:
+        func (callable): The function to be wrapped
+        v (str): The minimum required version in the format X.Y.Z
+    Returns:
+        wrapper (callable): The decorated function
+    """
+    def decorator(func: F) -> F:
+        @wraps(func)
+        def wrapper(obj, *args, **kwargs):
+            if isinstance(obj, BotMaestroSDK):
+                if obj.version is None:
+                    if obj.RAISE_NOT_CONNECTED:
+                        raise RuntimeError('Maestro version not available. Make sure to invoke login first.')
+                else:
+                    if version.parse(obj.version) < version.parse(v):
+                        message = f'''
+                        The method {func.__name__} is not available for your version of BotCity Maestro.
+                        Your version: {obj.version} - Required version: {v}. Please request an update.
+                        '''
+                        raise RuntimeError(message)
+            else:
+                raise NotImplementedError('since_version is only valid for BotMaestroSDK methods.')
+            return func(obj, *args, **kwargs)
+
+        return cast(F, wrapper)
+    return decorator
+
+
+class BotMaestroSDK:
+    _notified_disconnect = False
+    RAISE_NOT_CONNECTED = True
+    # More details about VERIFY_SSL_CERT here
+    # https://requests.readthedocs.io/en/latest/user/advanced/#ssl-cert-verification
+    VERIFY_SSL_CERT = True
 
     def __init__(self, server: Optional[str] = None, login: Optional[str] = None, key: Optional[str] = None):
         """
@@ -28,31 +127,102 @@ class BotMaestroSDK(BotMaestroSDKInterface):
         Attributes:
             access_token (str): The access token obtained via login.
         """
-        super().__init__(server=server, login=login, key=key)
+        self._server = None
+        self._login = login
+        self._key = key
+        self._access_token = None
+        self._task_id = 0
+        self._version = None
+        self._timeout = 30.0
+
+        self.server = server
 
     def _define_implementation(self):
         try:
             url = f'{self._server}/api/v2/maestro/version'
 
             with requests.get(url, verify=self.VERIFY_SSL_CERT) as req:
-                try:
-                    if req.status_code == 200:
-                        self._impl = v2.BotMaestroSDKV2(self.server, self._login, self._key, sdk=self)
-                        self._version = req.json()['version']
-                finally:
-                    if self._impl is None:
-                        self._impl = v1.BotMaestroSDKV1(self.server, self._login, self._key, sdk=self)
-                        self._version = "1.0.0"
-            self._impl.access_token = self.access_token
-            self._impl._login = self._login
+                req.raise_for_status()
+                self._version = req.json()['version']
         except Exception as ex:
             if self.RAISE_NOT_CONNECTED:
                 raise ex
-            self._impl = v2.BotMaestroSDKV2(self.server, self._login, self._key, sdk=self)
             self._version = "999.0.0"
-            self._impl._timeout = self.timeout
-            self._impl.access_token = self.access_token
-            self._impl._login = self._login
+
+    def _headers(self) -> Dict:
+        """The HTTP header for BotCity Maestro communication"""
+        return {
+            "Content-Type": "application/json",
+            "token": self.access_token,
+            "organization": self.organization
+        }
+
+    @classmethod
+    def from_sys_args(cls, default_server="", default_login="", default_key=""):
+        if len(sys.argv) >= 4:
+            maestro = cls()
+            organization = ""
+            try:
+                server, task_id, token, organization, *_ = sys.argv[1:]
+            except ValueError:
+                server, task_id, token, *_ = sys.argv[1:]
+            maestro.server = server
+            maestro.access_token = token
+            maestro.organization = organization
+            maestro.task_id = task_id
+        else:
+            maestro = cls(
+                server=default_server,
+                login=default_login,
+                key=default_key
+            )
+            if default_server:
+                maestro.login()
+        return maestro
+
+    @property
+    def server(self):
+        """The server address"""
+        return self._server
+
+    @server.setter
+    def server(self, server):
+        # Remove additional end /
+        if server and server[-1] == "/":
+            server = server[:-1]
+        self._server = server
+
+    @property
+    def access_token(self):
+        """The access token"""
+        return self._access_token
+
+    @access_token.setter
+    def access_token(self, token):
+        self._access_token = token
+
+    @property
+    def organization(self):
+        """The organization label"""
+        return self._login
+
+    @organization.setter
+    def organization(self, organization):
+        self._login = organization
+
+    @property
+    def task_id(self):
+        """The Current Task ID"""
+        return self._task_id
+
+    @task_id.setter
+    def task_id(self, task_id):
+        self._task_id = task_id
+
+    @property
+    def version(self):
+        """The BotCity Maestro Backend version"""
+        return self._version
 
     @property
     def timeout(self):
@@ -61,8 +231,6 @@ class BotMaestroSDK(BotMaestroSDKInterface):
     @timeout.setter
     def timeout(self, value):
         self._timeout = value
-        if self._impl:
-            self._impl._timeout = value
 
     def login(self, server: Optional[str] = None, login: Optional[str] = None, key: Optional[str] = None):
         """
@@ -89,10 +257,49 @@ class BotMaestroSDK(BotMaestroSDKInterface):
         self.logoff()
 
         self._define_implementation()
-        self._impl.login()
-        self.access_token = self._impl.access_token
+        url = f'{self._server}/api/v2/workspace/login'
+        data = {"login": self.organization, "key": self._key}
+        headers = {'Content-Type': 'application/json'}
 
-    @ensure_implementation()
+        with requests.post(url, data=json.dumps(data), headers=headers, timeout=self._timeout) as req:
+            if req.ok:
+                self.access_token = req.json()['accessToken']
+            else:
+                raise ValueError('Error during login. Server returned %d. %s' % (req.status_code, req.text))
+
+    def logoff(self):
+        """
+        Revoke the access token used to communicate with the BotMaestro portal.
+        """
+        self.access_token = None
+
+    @ensure_access_token(invoke=True)
+    def get_execution(self, task_id: Optional[str] = None) -> model.BotExecution:
+        """
+        Fetch the BotExecution object for a given task.
+
+        Args:
+            task_id (Optional[str], optional): The task ID. Defaults to None.
+
+        Returns:
+            model.BotExecution: The BotExecution information.
+        """
+        task_id = task_id or self.task_id
+
+        if not self.access_token and not self.RAISE_NOT_CONNECTED:
+            return model.BotExecution("", task_id, "", {})
+
+        if not task_id:
+            # If we are connected (access_token) or want to raise errors when disconnected
+            # we show the error, otherwise we are working offline and just want to ignore this
+            # but we will print a warning message for good measure
+            raise ValueError("A task ID must be informed either via the parameter or the class property.")
+
+        parameters = self.get_task(task_id).parameters
+
+        execution = model.BotExecution(self.server, task_id, self.access_token, parameters)
+        return execution
+
     @ensure_access_token()
     def alert(self, task_id: str, title: str, message: str, alert_type: model.AlertType) -> model.ServerMessage:
         """
@@ -107,9 +314,18 @@ class BotMaestroSDK(BotMaestroSDKInterface):
         Returns:
             Server response message. See [ServerMessage][botcity.maestro.model.ServerMessage]
         """
-        return self._impl.alert(task_id=task_id, title=title, message=message, alert_type=alert_type)
+        url = f'{self._server}/api/v2/alerts'
 
-    @ensure_implementation()
+        data = {"taskId": task_id, "title": title,
+                "message": message, "type": alert_type}
+
+        with requests.post(url, json=data, headers=self._headers(), timeout=self._timeout) as req:
+            if req.ok:
+                return model.ServerMessage.from_json(req.text)
+            else:
+
+                raise ValueError('Error during alert. %s', req.text)
+
     @ensure_access_token()
     def message(self, email: List[str], users: List[str], subject: str, body: str,
                 msg_type: model.MessageType, group: Optional[str] = None) -> model.ServerMessage:
@@ -127,10 +343,22 @@ class BotMaestroSDK(BotMaestroSDKInterface):
         Returns:
             Server response message. See [ServerMessage][botcity.maestro.model.ServerMessage]
         """
-        return self._impl.message(email=email, users=users, subject=subject, body=body, msg_type=msg_type,
-                                  group=group)
+        url = f'{self._server}/api/v2/message'
 
-    @ensure_implementation()
+        if not group:
+            group = ""
+
+        data = {"emails": email, "logins": users, "subject": subject, "body": body,
+                "type": msg_type, "group": group}
+        with requests.post(url, json=data, headers=self._headers(), timeout=self._timeout) as req:
+            if req.status_code != 200:
+                raise ValueError(
+                    'Error during message. Server returned %d. %s' %
+                    (req.status_code, req.json().get('message', ''))
+                )
+            payload = json.dumps({"message": req.text, "type": req.status_code})
+            return model.ServerMessage.from_json(payload=payload)
+
     @ensure_access_token()
     def create_task(self, activity_label: str, parameters: Dict[str, object],
                     test: bool = False, priority: int = 0, min_execution_date=None) -> model.AutomationTask:
@@ -147,10 +375,29 @@ class BotMaestroSDK(BotMaestroSDKInterface):
         Returns:
             Automation Task. See [AutomationTask][botcity.maestro.model.AutomationTask]
         """
-        return self._impl.create_task(activity_label=activity_label, parameters=parameters, test=test,
-                                      min_execution_date=min_execution_date, priority=priority)
+        url = f'{self._server}/api/v2/task'
 
-    @ensure_implementation()
+        data = {
+            "activityLabel": activity_label, "test": test, "parameters": parameters, "priority": priority,
+        }
+
+        if min_execution_date is not None:
+            if not isinstance(min_execution_date, datetime.datetime):
+                raise ValueError(f"Arg 'min_execution_date' is not datetime. Type {type(min_execution_date)}")
+            data["minExecutionDate"] = min_execution_date.isoformat()
+
+        headers = self._headers()
+        with requests.post(url, json=data, headers=headers, timeout=self._timeout) as req:
+            if req.ok:
+                return model.AutomationTask.from_json(req.text)
+            else:
+                try:
+                    message = 'Error during task create. Server returned %d. %s' % (
+                        req.status_code, req.json().get('message', ''))
+                except ValueError:
+                    message = 'Error during task create. Server returned %d. %s' % (req.status_code, req.text)
+                raise ValueError(message)
+
     @ensure_access_token()
     def finish_task(self, task_id: str, status: model.AutomationTaskFinishStatus,
                     message: str = "") -> model.ServerMessage:
@@ -166,9 +413,21 @@ class BotMaestroSDK(BotMaestroSDKInterface):
         Returns:
             Server response message. See [ServerMessage][botcity.maestro.model.ServerMessage]
         """
-        return self._impl.finish_task(task_id=task_id, status=status, message=message)
+        url = f'{self._server}/api/v2/task/{task_id}'
+        data = {"finishStatus": status, "finishMessage": message,
+                "state": "FINISHED"}
+        headers = self._headers()
+        with requests.post(url, json=data, headers=headers, timeout=self._timeout) as req:
+            if req.ok:
+                return model.ServerMessage.from_json(req.text)
+            else:
+                try:
+                    message = 'Error during task finish. Server returned %d. %s' % (
+                        req.status_code, req.json().get('message', ''))
+                except ValueError:
+                    message = 'Error during task finish. Server returned %d. %s' % (req.status_code, req.text)
+                raise ValueError(message)
 
-    @ensure_implementation()
     @ensure_access_token()
     def restart_task(self, task_id: str) -> model.ServerMessage:
         """
@@ -180,9 +439,20 @@ class BotMaestroSDK(BotMaestroSDKInterface):
         Returns:
             Server response message. See [ServerMessage][botcity.maestro.model.ServerMessage]
         """
-        return self._impl.restart_task(task_id=task_id)
+        url = f'{self._server}/api/v2/task/{task_id}'
+        data = {"state": "START"}
+        headers = self._headers()
+        with requests.post(url, json=data, headers=headers, timeout=self._timeout) as req:
+            if req.ok:
+                return model.ServerMessage.from_json(req.text)
+            else:
+                try:
+                    message = 'Error during task restart. Server returned %d. %s' % (
+                        req.status_code, req.json().get('message', ''))
+                except ValueError:
+                    message = 'Error during task restart. Server returned %d. %s' % (req.status_code, req.text)
+                raise ValueError(message)
 
-    @ensure_implementation()
     @ensure_access_token()
     def get_task(self, task_id: str) -> model.AutomationTask:
         """
@@ -194,9 +464,20 @@ class BotMaestroSDK(BotMaestroSDKInterface):
         Returns:
             Automation Task. See [AutomationTask][botcity.maestro.model.AutomationTask]
         """
-        return self._impl.get_task(task_id=task_id)
+        url = f'{self._server}/api/v2/task/{task_id}'
 
-    @ensure_implementation()
+        with requests.get(url, headers=self._headers(), timeout=self._timeout) as req:
+            if req.ok:
+                payload = req.text
+                return model.AutomationTask.from_json(payload)
+            else:
+                try:
+                    message = 'Error during task get. Server returned %d. %s' % (
+                        req.status_code, req.json().get('message', ''))
+                except ValueError:
+                    message = 'Error during task get. Server returned %d. %s' % (req.status_code, req.text)
+                raise ValueError(message)
+
     @since_version("2.0.0")
     @ensure_access_token()
     def interrupt_task(self, task_id: str) -> model.ServerMessage:
@@ -209,9 +490,20 @@ class BotMaestroSDK(BotMaestroSDKInterface):
         Returns:
             Server response message. See [ServerMessage][botcity.maestro.model.ServerMessage]
         """
-        return self._impl.interrupt_task(task_id=task_id)
+        url = f'{self._server}/api/v2/task/{task_id}'
+        data = {"interrupted": True}
+        headers = self._headers()
+        with requests.post(url, json=data, headers=headers, timeout=self._timeout) as req:
+            if req.ok:
+                return model.ServerMessage.from_json(req.text)
+            else:
+                try:
+                    message = 'Error during task finish. Server returned %d. %s' % (
+                        req.status_code, req.json().get('message', ''))
+                except ValueError:
+                    message = 'Error during task finish. Server returned %d. %s' % (req.status_code, req.text)
+                raise ValueError(message)
 
-    @ensure_implementation()
     @ensure_access_token()
     def new_log(self, activity_label: str, columns: List[model.Column]) -> model.ServerMessage:
         """
@@ -224,9 +516,22 @@ class BotMaestroSDK(BotMaestroSDKInterface):
         Returns:
             Server response message. See [ServerMessage][botcity.maestro.model.ServerMessage]
         """
-        return self._impl.new_log(activity_label=activity_label, columns=columns)
+        url = f'{self._server}/api/v2/log'
 
-    @ensure_implementation()
+        cols = [asdict(c) for c in columns]
+
+        data = {"activityLabel": activity_label, "columns": cols, 'organizationLabel': self.organization}
+        with requests.post(url, json=data, headers=self._headers(), timeout=self._timeout) as req:
+            if req.ok:
+                return model.ServerMessage.from_json(req.text)
+            else:
+                try:
+                    message = 'Error during new log. Server returned %d. %s' % (
+                        req.status_code, req.json().get('message', ''))
+                except ValueError:
+                    message = 'Error during new log. Server returned %d. %s' % (req.status_code, req.text)
+                raise ValueError(message)
+
     @ensure_access_token()
     def new_log_entry(self, activity_label: str, values: Dict[str, object]) -> model.ServerMessage:
         """
@@ -239,9 +544,19 @@ class BotMaestroSDK(BotMaestroSDKInterface):
         Returns:
             Server response message. See [ServerMessage][botcity.maestro.model.ServerMessage]
         """
-        return self._impl.new_log_entry(activity_label=activity_label, values=values)
+        url = f'{self._server}/api/v2/log/{activity_label}/entry'
 
-    @ensure_implementation()
+        with requests.post(url, json=values, headers=self._headers(), timeout=self._timeout) as req:
+            if req.status_code != 200:
+                try:
+                    message = 'Error during new log entry. Server returned %d. %s' % (
+                        req.status_code, req.json().get('message', ''))
+                except ValueError:
+                    message = 'Error during new log entry. Server returned %d. %s' % (req.status_code, req.text)
+                raise ValueError(message)
+            payload = json.dumps({"message": req.text, "type": req.status_code})
+            return model.ServerMessage.from_json(payload=payload)
+
     @ensure_access_token()
     def get_log(self, activity_label: str, date: Optional[str] = "") -> List[Dict[str, object]]:
         """
@@ -255,9 +570,48 @@ class BotMaestroSDK(BotMaestroSDKInterface):
             Log entry list. Each element in the list is a dictionary in which keys are Column names and values are
             the column value.
         """
-        return self._impl.get_log(activity_label=activity_label, date=date)
+        url = f'{self._server}/api/v2/log/{activity_label}'
 
-    @ensure_implementation()
+        days = 365  # 1 year is enough
+        if date:
+            days = (datetime.datetime.now()-datetime.datetime.strptime(date, "%d/%m/%Y")).days + 1
+
+        with requests.get(url, headers=self._headers(), timeout=self._timeout) as req:
+            if req.ok:
+                log = req.json()
+                columns = log.get('columns')
+                if not columns:
+                    raise ValueError('Malformed log. No columns available.')
+                names_for_labels = {c['label']: c['name'] for c in columns}
+                url = f'{self._server}/api/v2/log/{activity_label}/entry-list'
+
+                data = {"days": days}
+                with requests.get(url, params=data, headers=self._headers(), timeout=self._timeout) as entry_req:
+                    if entry_req.ok:
+                        log_data = []
+                        for en in entry_req.json():
+                            cols = en['columns']
+                            d = dict()
+                            for label, name in names_for_labels.items():
+                                d[name] = cols[label]
+                            log_data.append(d)
+                        return log_data
+                    else:
+                        try:
+                            message = 'Error during log entry read. Server returned %d. %s' % (
+                                entry_req.status_code, entry_req.json().get('message', ''))
+                        except ValueError:
+                            message = 'Error during log entry read. Server returned %d. %s' % (
+                                entry_req.status_code, entry_req.text)
+                        raise ValueError(message)
+            else:
+                try:
+                    message = 'Error during log read. Server returned %d. %s' % (
+                        req.status_code, req.json().get('message', ''))
+                except ValueError:
+                    message = 'Error during log read. Server returned %d. %s' % (req.status_code, req.text)
+                raise ValueError(message)
+
     @ensure_access_token()
     def delete_log(self, activity_label: str) -> model.ServerMessage:
         """
@@ -270,9 +624,21 @@ class BotMaestroSDK(BotMaestroSDKInterface):
             Log entry list. Each element in the list is a dictionary in which keys are Column names and values are
             the column value.
         """
-        return self._impl.delete_log(activity_label=activity_label)
+        # date  a partir desta data
+        # date em branco eh tudo
+        url = f'{self._server}/api/v2/log/{activity_label}'
 
-    @ensure_implementation()
+        with requests.delete(url, headers=self._headers(), timeout=self._timeout) as req:
+            if req.status_code != 200:
+                try:
+                    message = 'Error during log delete. Server returned %d. %s' % (
+                        req.status_code, req.json().get('message', ''))
+                except ValueError:
+                    message = 'Error during log delete. Server returned %d. %s' % (req.status_code, req.text)
+                raise ValueError(message)
+            payload = json.dumps({"message": req.text, "type": req.status_code})
+            return model.ServerMessage.from_json(payload=payload)
+
     @ensure_access_token()
     def post_artifact(self, task_id: int, artifact_name: str, filepath: str) -> model.ServerMessage:
         """
@@ -286,20 +652,80 @@ class BotMaestroSDK(BotMaestroSDKInterface):
         Returns:
             Server response message. See [ServerMessage][botcity.maestro.model.ServerMessage]
         """
-        return self._impl.post_artifact(task_id=task_id, artifact_name=artifact_name, filepath=filepath)
+        artifact_id = self._create_artifact(task_id=task_id, name=artifact_name, filename=artifact_name)
+        url = f'{self._server}/api/v2/artifact/log/{json.loads(artifact_id.payload)["id"]}'
 
-    @ensure_implementation()
+        with open(filepath, 'rb') as f:
+            data = MultipartEncoder(
+                fields={'file': (artifact_name, f)}
+            )
+            headers = {**self._headers(), 'Content-Type': data.content_type}
+            with requests.post(url, data=data, headers=headers, timeout=self._timeout) as req:
+                if req.ok:
+                    return artifact_id
+                else:
+                    try:
+                        message = 'Error during artifact posting. Server returned %d. %s' % (
+                            req.status_code, req.json().get('message', ''))
+                    except ValueError:
+                        message = 'Error during artifact posting. Server returned %d. %s' % (
+                            req.status_code, req.text)
+                    raise ValueError(message)
+
     @ensure_access_token()
-    def list_artifacts(self) -> List[model.Artifact]:
+    def _create_artifact(self, task_id: int, name: str, filename: str) -> model.ServerMessage:
+        """
+        Creates a new artifact
+
+        Args:
+            task_id: The task unique identifier.
+            name: The name of the artifact to be displayed on the portal.
+            filename: The file to be uploaded.
+
+        Returns:
+            Server response message. See [ServerMessage][botcity.maestro.model.ServerMessage]
+        """
+        url = f'{self._server}/api/v2/artifact'
+        data = {'taskId': task_id, 'name': name, 'filename': filename}
+        with requests.post(url, json=data, headers=self._headers(), timeout=self._timeout) as req:
+            if req.ok:
+                return model.ServerMessage.from_json(req.text)
+            else:
+                try:
+                    message = 'Error during new log entry. Server returned %d. %s' % (
+                        req.status_code, req.json().get('message', ''))
+                except ValueError:
+                    message = 'Error during new log entry. Server returned %d. %s' % (req.status_code, req.text)
+                raise ValueError(message)
+
+    @ensure_access_token()
+    def list_artifacts(self, days: int = 7) -> List[model.Artifact]:
         """
         List all artifacts available for the organization.
 
         Returns:
             List of artifacts. See [Artifact][botcity.maestro.model.Artifact]
         """
-        return self._impl.list_artifacts()
+        url = f'{self._server}/api/v2/artifact?size=5&page=0&sort=dateCreation,desc&days={days}'
 
-    @ensure_implementation()
+        with requests.get(url, headers=self._headers(), timeout=self._timeout) as req:
+            if req.ok:
+                content = req.json()['content']
+                response = [model.Artifact.from_dict(a) for a in content]
+                for page in range(1, req.json()['totalPages']):
+                    url = f'{self._server}/api/v2/artifact?size=5&page={page}&sort=dateCreation,desc&days={days}'
+                    with requests.get(url, headers=self._headers(), timeout=self._timeout) as req:
+                        content = req.json()['content']
+                        response.extend([model.Artifact.from_dict(a) for a in content])
+                return response
+            else:
+                try:
+                    message = 'Error during artifact listing. Server returned %d. %s' % (
+                        req.status_code, req.json().get('message', ''))
+                except ValueError:
+                    message = 'Error during artifact listing. Server returned %d. %s' % (req.status_code, req.text)
+                raise ValueError(message)
+
     @ensure_access_token()
     def get_artifact(self, artifact_id: int) -> Tuple[str, bytes]:
         """
@@ -311,9 +737,26 @@ class BotMaestroSDK(BotMaestroSDKInterface):
         Returns:
             Tuple containing the artifact name and an array of bytes which are the binary content of the artifact.
         """
-        return self._impl.get_artifact(artifact_id=artifact_id)
+        url = f'{self._server}/api/v2/artifact/{artifact_id}'
 
-    @ensure_implementation()
+        with requests.get(url, headers=self._headers(), timeout=self._timeout) as req:
+            if req.ok:
+                payload = req.json()
+                filename = payload['fileName']
+
+                url = f'{self.server}/api/v2/artifact/{artifact_id}/file'
+                with requests.get(url, headers=self._headers(), timeout=self._timeout) as req_file:
+                    file_content = req_file.content
+
+                return filename, file_content
+            else:
+                try:
+                    message = 'Error during artifact get. Server returned %d. %s' % (
+                        req.status_code, req.json().get('message', ''))
+                except ValueError:
+                    message = 'Error during artifact get. Server returned %d. %s' % (req.status_code, req.text)
+                raise ValueError(message)
+
     @since_version("2.0.0")
     @ensure_access_token()
     def error(self, task_id: int, exception: Exception, screenshot: Optional[str] = None,
@@ -333,9 +776,137 @@ class BotMaestroSDK(BotMaestroSDKInterface):
         Raises:
             ValueError: If the request fails, a ValueError exception is raised.
         """
-        return self._impl.error(task_id, exception, screenshot, attachments, tags)
+        url = f'{self._server}/api/v2/error'
+        trace = " ".join(traceback.format_exception(type(exception), exception, exception.__traceback__))
 
-    @ensure_implementation()
+        if not tags:
+            tags = dict()
+
+        default_tags = self._get_default_error_tags()
+        default_tags.update(tags)
+        tags = default_tags
+
+        data = {'taskId': task_id, 'type': exception.__class__.__name__, 'message': str(exception),
+                'stackTrace': trace, 'language': 'PYTHON', 'tags': tags}
+
+        response = None
+        with requests.post(url, json=data, headers=self._headers(), timeout=self._timeout) as req:
+            if req.status_code == 201:
+                response = req.json()
+            else:
+                try:
+                    message = 'Error during new error entry. Server returned %d. %s' % (
+                        req.status_code, req.json().get('message', ''))
+                except ValueError:
+                    message = 'Error during new error entry. Server returned %d. %s' % (req.status_code, req.text)
+                raise ValueError(message)
+
+        if screenshot:
+            self._create_screenshot(error_id=response.get('id'), filepath=screenshot)
+
+        # pip list
+        packages = [(dist.name.lower(), dist.version) for dist in importlib_metadata.distributions()]
+        packages.sort(key=lambda x: x[0])  # type: ignore
+        buffer = StringIO()
+        buffer.writelines([f"{name}=={version}{os.linesep}" for name, version in packages])
+        buffer.flush()
+        self._create_attachment(
+            error_id=response.get('id'),
+            filename="piplist.txt",
+            buffer=buffer
+        )
+        buffer.close()
+
+        if attachments:
+            for attachment in attachments:
+                filepath = os.path.expandvars(os.path.expanduser(attachment))
+                with open(filepath, 'rb') as f:
+                    self._create_attachment(
+                        error_id=response.get('id'),
+                        filename=Path(filepath).name,
+                        buffer=f
+                    )
+
+        return response
+
+    def _get_default_error_tags(self) -> Dict:
+        """Generates a dictionarty with useful tags about the system for the error method
+        """
+        tags = dict()
+        try:
+            tags["user_name"] = os.getlogin()
+        except Exception:
+            tags["user_name"] = ""
+        tags["host_name"] = platform.node()
+        tags["os_name"] = platform.system()
+
+        os_version = platform.version()
+        if platform.system() == "Linux":
+            os_version = " ".join(distro.linux_distribution())
+        elif platform.system() == "Darwin":
+            os_version = platform.mac_ver()[0]
+
+        tags["os_version"] = os_version
+        tags["python_version"] = platform.python_version()
+
+        return tags
+
+    def _create_screenshot(self, error_id: int, filepath: str) -> None:
+        """
+        Creates a new screenshot in error
+
+        Args:
+            error_id: The error unique identifier.
+            filepath: File path for screenshot
+        Returns:
+            None
+        """
+        url_screenshot = f'{self._server}/api/v2/error/{error_id}/screenshot'
+        filepath = os.path.expandvars(os.path.expanduser(filepath))
+
+        with open(filepath, 'rb') as f:
+            data_screenshot = MultipartEncoder(
+                fields={'file': (Path(filepath).name, f)}
+            )
+            headers = self._headers()
+            headers['Content-Type'] = data_screenshot.content_type
+
+            with requests.post(url_screenshot, data=data_screenshot, headers=headers, timeout=self._timeout) as req:
+                if not req.ok:
+                    try:
+                        message = 'Error during new log entry. Server returned %d. %s' % (
+                            req.status_code, req.json().get('message', ''))
+                    except ValueError:
+                        message = 'Error during new log entry. Server returned %d. %s' % (
+                            req.status_code, req.text)
+                    raise ValueError(message)
+
+    def _create_attachment(self, error_id: int, filename: str, buffer: IOBase):
+        """
+        Creates a new attachment in error
+
+        Args:
+            error_id (int): The error unique identifier.
+            filename (str): The file name to be displayed.
+            buffer (IOBase): The file handler buffer.
+        """
+        url_attachments = f'{self._server}/api/v2/error/{error_id}/attachments'
+
+        file = MultipartEncoder(
+            fields={'file': (filename, buffer)}
+        )
+        headers = self._headers()
+        headers['Content-Type'] = file.content_type
+        with requests.post(url_attachments, data=file, headers=headers, timeout=self._timeout) as req:
+            if not req.ok:
+                try:
+                    message = 'Error during new log entry. Server returned %d. %s' % (
+                        req.status_code, req.json().get('message', ''))
+                except ValueError:
+                    message = 'Error during new log entry. Server returned %d. %s' % (
+                        req.status_code, req.text)
+                raise ValueError(message)
+
     @since_version("2.0.0")
     @ensure_access_token()
     def get_credential(self, label: str, key: str) -> str:
@@ -348,9 +919,19 @@ class BotMaestroSDK(BotMaestroSDKInterface):
         Returns:
             Key value that was requested
         """
-        return self._impl.get_credential(label, key)
+        url = f'{self._server}/api/v2/credential/{label}/key/{key}'
 
-    @ensure_implementation()
+        with requests.get(url, headers=self._headers(), timeout=self._timeout) as req:
+            if req.ok:
+                return str(req.text)
+            else:
+                try:
+                    message = 'Error during log read. Server returned %d. %s' % (
+                        req.status_code, req.json().get('message', ''))
+                except ValueError:
+                    message = 'Error during log read. Server returned %d. %s' % (req.status_code, req.text)
+                raise ValueError(message)
+
     @since_version("2.0.0")
     @ensure_access_token()
     def create_credential(self, label: str, key: str, value):
@@ -361,9 +942,53 @@ class BotMaestroSDK(BotMaestroSDKInterface):
             key: Key name within the credential set
             value: Credential value
         """
-        return self._impl.create_credential(label, key, value)
+        credential = self._get_credential_by_label(label=label)
 
-    @ensure_implementation()
+        if credential is None:
+            response = self._create_credential_by_label(label=label, key=key, value=value)
+            if response is None:
+                raise ValueError('Error during create credential.')
+            return response.to_json()
+        data = {
+            'key': key,
+            'value': value
+        }
+        url = f'{self._server}/api/v2/credential/{label}/key'
+        with requests.post(url, json=data, headers=self._headers(), timeout=self._timeout) as req:
+            if not req.ok:
+                req.raise_for_status()
+
+    def _get_credential_by_label(self, label):
+        """
+        Get dict in key inside credentials
+        Args:
+            label: Credential set name
+        Returns:
+            Credential dict
+        """
+        url = f'{self._server}/api/v2/credential/{label}'
+
+        with requests.get(url, headers=self._headers(), timeout=self._timeout) as req:
+            if req.ok:
+                return model.ServerMessage.from_json(req.text)
+            else:
+                return None
+
+    def _create_credential_by_label(self, label: str, key: str, value):
+        data = {
+            'label': label,
+            'secrets': [
+                {'key': key, 'value': value, 'valid': True}
+            ]
+        }
+        url = f'{self._server}/api/v2/credential'
+
+        with requests.post(url, json=data, headers=self._headers(), timeout=self._timeout) as req:
+            if req.ok:
+                return model.ServerMessage.from_json(req.text)
+            else:
+                return None
+
     @since_version("3.0.2")
     @ensure_access_token()
     def create_datapool(self, pool) -> DataPool:
@@ -376,10 +1001,14 @@ class BotMaestroSDK(BotMaestroSDKInterface):
         Returns:
             DataPool instance. See [DataPool][ [DataPool][botcity.maestro.datapool.DataPool] instance.
         """
-        new_pool = self._impl.create_datapool(pool=pool)
-        return new_pool
+        url = f'{self.server}/api/v2/datapool'
+        pool.maestro = self
+        with requests.post(url, data=json.dumps(pool.to_dict()), headers=self._headers(),
+                           timeout=self._timeout) as req:
+            if req.ok:
+                return pool
+            req.raise_for_status()
 
-    @ensure_implementation()
     @since_version("3.0.2")
     @ensure_access_token()
     def get_datapool(self, label: str) -> DataPool:
@@ -392,5 +1021,9 @@ class BotMaestroSDK(BotMaestroSDKInterface):
         Returns:
             DataPool instance. See [DataPool][ [DataPool][botcity.maestro.datapool.DataPool] instance.
         """
-        pool = self._impl.get_datapool(label=label)
-        return pool
+        url = f'{self._server}/api/v2/datapool/{label}'
+
+        with requests.get(url, headers=self._headers(), timeout=self._timeout) as req:
+            if req.ok:
+                return DataPool.from_json(payload=req.content, maestro=self)
+            req.raise_for_status()
